Move UMA2 resources to their own file
